Corrige cadastro de equipamento

+++ src/main/java/br/com/audiovisual/dao/InterfaceDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface InterfaceDAO<T> {
 	public void salvar(T obj) throws SQLException;
 	public void atualizar(T obj) throws SQLException;
-	public void excluir(int id) throws SQLException;
+	public void excluir(Long id) throws SQLException;
 	public T consultarById(int id);
 	public List<T> consultarTodos() throws SQLException;
 }
